requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
altair==5.5.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
inference-sdk==0.55.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.5.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polars==1.33.1
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
supervision==0.26.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.201
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
